requirements.txt
Django>=5.0
google-generativeai
python-dotenv
Pillow
replicate
requests
markdown
